meta-openembedded: waf-samba.bbclass: add cross-answers option and files
--cross-execute uses qemu to run cross-compiled binaries,
it works fine for qemu supported targets, but it definitely
will fail if the target is not supported by qemu. So we need
to use the cross-answers option.

Added three methods and set 'answer' as default:

1. answers:
   Only --cross-answers - try the cross-answers file, and if
   there's no corresponding answer, add to the file and mark
   the configure process as unfinished.
2. exec:
   Only --cross-execute - get the answer from cross-execute,
   an emulator (qemu) is used to run cross-compiled binaries.
3. both:
   Try the cross-answers file first, and if there is no
   corresponding answer, use cross-execute to get an answer,
   and add that answer to the file.

The cross files in are all generated by qemu when the 'both'
mode is enabled, and are placed in a common directory
LAYERDIR/files/waf-cross-answers which is set as
WAF_CROSS_ANSWERS_PATH in layer.conf of meta-oe,
so that we don't have to put copies of them for each
package (there are 5 packages inerit waf-samba for now)

(From meta-openembedded rev: fe4106ca0d81fb0dd30c82a2c4a45c38f9694ad6)

Signed-off-by: Jackie Huang <jackie.huang@windriver.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>
Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>