meta-openembedded: mongodb: 2.6.0 -> 3.3.0 and remove from pnblacklist
Since building mongodb was impossible currently, I updated the recipe
and got it working on arm >= v6, x86 and x86_64. Unfortunately, it fails
building with system-boost for the current version 1.60. The corresponding
ticket SERVER-17294 of mongodb jira is still in progress. So we use the
built-in version of boost.

(From meta-openembedded rev: 0c024ec106373fc2fd2d07e596a295809d16b4bd)

Signed-off-by: Sven Ebenfeld <sven.ebenfeld@gmail.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>