toybox_%.bbappend: add dummy mesg command
The default .profile calls "mesg" to tighten the security
of the interactive terminal. While this is useful, it is not
a use cases that is supported by "production" Ostro OS images
and thus not important.

Toybox currently does not provide a mesg implementation.
To avoid the error when the command is missing, a fake shell
script gets installed instead.

Linking to /bin/true would be simpler, but does not work because true
symlinks to toybox, leading to invoking toybox under the "mesg" name
which then does not cause toybox to execute the "true" (pun intended)
command...

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>