meta-swupd: swupd-image: remove the shared pseudo database after running swupd
We need to use a shared pseudo database whilst all of the virtual
images are created and the swupd post-processing is performed so
that the permissions are correct across recipes, however we also
need to be careful about the database's lifecycle to ensure that
operations run outside of pseudo control don't cause the database
to become invalidated.

As the do_swupd_image task only runs for the base image *and* runs
after all of the virtual images have been created we can safely
clean up the database once it's finished — add a task to do so.

(From meta-swupd rev: 0c197c504d3ff267390a9ddbe32be3f1f4f1a960)

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>