initramfs-framework: avoid complex expr
Toybox does not currently implement expr and even if it did, whether
regular expression support would be added right away is
unclear. Replacing the expr command with cut allows using the debug
features in an initramfs which uses toybox instead of busybox.

The "cut" command is also (somewhat subjectively) easier to read.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>