ostro.conf: remove systemd-serialgetty
systemd-serialgetty may have some usage in other distros or platforms
supported by OE-core (see "[OE-core] systemd-serial-getty obsolete?"),
but in Ostro OS we do not need and do not want it because
systemd-getty-generator works for us.

It gets removed in ostro.conf because a systemd_%.bbappend would be
overkill for this one-line change and this package config change can
also be combined nicely with blacklisting the systemd-serialgetty
recipe to prevent accidentally pulling it in.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>