DOC: document the available compression methods
The recommended compression types get mentioned explicitly.

For the others it would be nice to have a link, but the upstream
support only seems to be documented in the source code. As this is for
people building images, pointing them to the .bbclass seems
sufficient.

Replicating the current list from the .bbclass was intentionally
avoided because the list might change over time.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>