image creation: fix dependency calculation when using conversion chaining
When using conversion chaining (for example example: .dsk.vdi.xz),
the imagetypes_getdepends() did not properly detect all compression
commands (thus skipping the corresponding COMPRESS_DEPENDS) and
the base type, leading to missing dependencies of the image's do_rootfs
task.

This is not a big problem in practice because in those cases where
conversion chaining is useful (as in the example above), the missing
dependency is qemu-native, which typically gets built anyway.

The solution moves imagetypes_getdepends() to image.bbclass because that
is the only place where it gets used and rewrites it so that it and
generating the image commands use a new common utility method, image_split_type().

imagetypes_getdepends() had hard-coded special treatment for certain
image types. This gets replaced with setting the normal IMAGE_DEPENDS
variables for these types.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>