stateless: introduce path check
Moving files like /etc/ssl/certs/ca-certificates.crt from well-known
locations to something new only works if all software relying on the
former gets updated or reconfigured accordingly.

Checking for this is error-prone, in particular when we consider that
new software may be added in the future. This new check greps all
packaged files for strings which should not be in use anymore, like
"/etc/ssl/certs".

It still possible that the path gets composed from individual
strings, in which case the test will not find the problem, but typical
usages (as in curl, openssl, an Perl) are found, so the test is
useful.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>