meta-swupd: swupd images: support creating images that contain os-core + additional bundles
To keep the "os-core" bundle small, the base image should be as
minimal as possible. It might not even allow logging in. In such a
setup, actual images have to have additional software installed, which
in a swupd-based image implies installing additional bundles.

The new SWUPD_IMAGES variable can be used to request the creation of
additional virtual image recipes which, if building them gets
explicitly requested, create a suitable rootfs by copying from the
mega image rootfs and then applying the normal image creation.

In contrast to the internal bundle-<baseimage>-<bundle> recipes, the
SWUPD_IMAGES are user-visible, so they are named
<baseimage>-<imageextension> where <imageextension> is the name given
in SWUPD_IMAGES.

To make this work, some changes are required in the code:
- the special swupd methods need to ignore the virtual image
  recipes which produce images
- MEGA_IMAGE_ROOTFS needs to be set for all image recipes
- /usr/share/clear/bundles is populated consistently in all
  rootfs directories (including os-core and also for the mega
  rootfs)
- manipulating the rootfs must be split into code which sets
  up the clear/bundles directory (can and should be in do_rootfs)
  and copying files (should better be in do_image, because it needs
  to replace work done there) and the latter must be available
  as an utility function because the new virtual image recipes
  replace the entire do_image with it
- the shared pseudo database can no longer be removed, because the
  image creation also depends on it, and it is unknown when that is
  done (might even be in a separate, future bitbake invocation)

The "swupdimage.bbclass" name is consistent with
"swupdbundle.bbclass", but can easily be confused with
"swupd-image.bbclass". The goal is to rename "swupd-image.bbclass" to
just "swupd.bbclass".

(From meta-swupd rev: 42f51d68b27d329bcb54090e42ecd276f6e0f25c)

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>