ostro: set OS_VERSION for swupd
The bundle information created by meta-swupd and the VERSION_ID used
in the os-release file must match, and therefore the value must be set
at the distro level. OS_VERSION is the variable used by meta-swupd,
so that is what is getting used in all places.

As before, the default OS_VERSION is derived from the integer part of
the BUILD_ID set in the Jenkins CI by multiplying it by 10 (aka
appending a string containing "0"). However, now the resulting value
of OS_VERSION can be overridden in the local build configuration or
via environment variables.

Also there's a very simplistic attempt to turn the default BUILD_ID =
${DATETIME} into a suitable OS_VERSION. It's better to set OS_VERSION
manually when building locally. This needs some more thoughts, in
particular because meta-swupd expects the value to
increase. Rebuilding with the same value currently does not update the
deployed bundle information.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>