ostro-image.bb: enable swupd and re-configure images
When the "swupd" image feature gets activated (off by default now),
the utility classes from meta-swupd are used to
a) generate swupd bundles
b) generate ostro-image-<foobar> images which consist of the core-os
   bundle plus additional, pre-installed bundles and have the swupd
   client installed.

To demonstrate how this could work, "ostro-image-minimal" and
"ostro-image-all" are pre-defined when building with swupd
enabled. The actual image names and content are to be defined.

When building without swupd, the traditional image variant support is
still enabled and pre-configured to build "ostro-image-minimal" and
"ostro-image-dev".

Building "ostro-image" itself works in both cases, but is resulting in
very minimal images, so that is only suitable with additional software
added to it.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>