ostro-image.bb: clean up and fix image features and bundle definitions
tools-profile was not used and does not compile in Ostro OS when
trying to use it, therefore it gets removed.

To simplify the bundles definition, image features get split into
those that are safe to use for bundles and those that are not because
they change configuration.

The default bundles then map 1:1 to those features which add packages.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>