images: re-enable ostro-image-dev
Although not the main use case, on-target compilation is still
useful in some cases and thus an image with development files
installed and the corresponding bundles are needed.

Based on the new support for -dev bundles, the re-enabled
ostro-image-dev replaces ostro-image-qa as the image which has
most software pre-installed and gets tested.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>