ostro: make BUILD_ID in os-release file CI specific
Including BUILD_ID only makes sense in the CI, because during local
builds it is set to a basically unpredictable value (${DATETIME}) and
does not get updated in os-release during later builds because of the
implications (constant rebuilds, inconsistencies between different
machines).

This change also documents what the BUILD_ID in our CI system is.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>