swupd: configure OS_VERSION differently in CI and local builds
BUILD_ID is only useful in the CI. Setting OS_VERSION based on it
therefore belongs into ostroproject-ci.inc.

During a local build, we can explicitly use ${DATETIME} to have a
reasonable default if developers do not set OS_VERSION manually. The
default is to use a modified OS_VERSION if something else triggers an
image build, but a OS_VERSION change itself won't trigger an image and
bundle build.

If a rebuild is desired without other changes, a developer can enforce
it by setting VERSION_ID in local.conf.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>