Ostro images: support images with and without swupd enabled in the same configuration
This makes it possible to build both in the CI system and simplifies
usage during local builds, because developers can switch from one kind
of image to another without having to change their local.conf.

The new ostro-image-swupd.bb is the base image for the virtual
ostro-image-swupd-reference and ostro-image-swupd-dev where swupd is
enabled. ostro-image-swupd does not produce image files, because those
would not be useful (too minimal).

Most of the code from ostro-image.bb was moved to ostro-image.bbclass
to allow sharing it with ostro-image-swupd.bb. Renaming ostro-image.bb
to something like ostro-image-swupd.bb would make sense, but cannot be
done because code elsewhere (for example, in meta-iotqa) still depends
on an image recipe called "ostro-image".

The "ostro-image" base name is currently reserved by the CI testing
system for the image which is to be used for automated
testing. Therefore and for the sake of consistency with the intended
naming, IMAGE_BASENAME gets overridden when building "ostro-image".
"ostro-image-noswupd" is also possible because of an alias set
via PROVIDES.

The CI gets configured to build ostro-image-noswupd,
ostro-image-swupd-reference and ostro-image-swupd-dev. CI sanity
testing happens on ostro-image-swupd-dev.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>