os-release: better tracking of BUILD_ID and and VERSION_ID
Recording volatile information in the os-release package is
conceptually problematic: either the package gets updated during each
build and thus causes constant image rebuilds, or it does not get
updated and then old, incorrect information gets copied into images.

In Ostro OS, we do not have to keep packaged files unmodified during
image creation, so the solution is to record static strings in the
os-release package and replace that with the actual, current values
during image creation.

This now works in the CI and in local builds, therefore recording
BUILD_ID can be enabled for both.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>