test-iot: eliminate EXTRA_IMAGEDEPENDS
Building additional test binaries through an image's
EXTRA_IMAGEDEPENDS always has been a hack. It was difficult
to prevent building them when a developer just wants to
build images, without testing them.

Now the approach with depending on the "qatests" image feature no
longer works, because in swupd-enabled builds none of the involved
images has that feature anymore. The ones that are getting tested
only install the on-device tests.

A cleaner solution is to extend the relevant task dependencies
directly, without the intermediate dependency through
"ostro-image:do_build" -> "${EXTRA_IMAGEDEPENDS}".

This is still not the final solution: instead of hard-coding
IOTQA_TESTIMAGEDEPENDS in test-iot.bbclass, the actual package list
should be configured elsewhere with a mapping from image name to
required packages for tests enabled on that image.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>