ostro-image.bbclass: make do_swupd_output visible in the console
The normal log level only puts progress messages into the log file
where is is not visible unless developers know where to look. In a CI
system the output is not accessible at all.

Using bbplain for progress messages is a bit unusual and thus not done
by default in meta-swupd, but okay (they just do not get formatted as
NOTES). A dedicated logging level for progress notifications would be
better, but does not exist in bitbake at the moment.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>