ostro-sanity.bbclass: add do_image QA check for dangling symlinks
Dangling symlinks have been a problem in particular in swupd-based
images, because it can happen that the actual file that gets pointed
to is only part of a bundle, while the link is part of the os-core.

This new check covers this. It gets applied to all images, whether
they are based on swupd or not, because we also do not want dangling
symlinks in non-swupd images.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>