ostro-image.bb: avoid task hash mismatch
A recent change in bitbake now triggers "Taskhash mismatch" errors for
do_rootfs when building Ostro images locally. CI builds are not affected
because BUILD_ID is set differently.

Ignoring the BUILD_ID -> DATETIME dependency looks like the right
solution, because conceptually we want the new value to be used when
it gets used, but we do not want a change in it to trigger a rebuild.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>