image-buildinfo.bbclass: support stateless distros
In a stateless distro, /etc/ cannot be used to store the "build"
file. The mechanism in Ostro OS which makes a normal rootfs
"stateless" also does not work in this case because it runs before the
code which creates /etc/build.

Therefore the class itself has to be parameterized such that the
actual location can be chosen by the distro. The default is unchanged
and still /etc/build.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>