stateless: import Clear Linux libpam patches
Required for setting root password when /etc is empty, because
otherwise PAM's "is changing the password allowed" check fails,
leading to a "permission denied" error before the password prompt.

Clear Linux also has a
0001-pam_securetty-Do-not-report-non-fatal-documented-beh.patch which
avoids a warning message when the /etc/securetty file is missing, but
in our version of libpam that code is different (patch does not apply)
and we do not trigger the warning, so the patch is not getting
imported.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>