motd: consistent handling in ssh and local logins
ssh logins did not show the motd and thus our "this is a development
image" warning. We want this to be shown for all interactive login
sessions, so now pam_motd.so is activated in the "common-session" file
and not just in the "login" file used for local logins.

The code is written so that it still works if PAM gets disabled and
the files do not get installed at all. It is also careful to not add
pam_motd.so twice and works when it also gets removed upstream
(i.e. it avoids "must update meta-ostro in sync with OE-core" issues).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>