ostro-image.bbclass: prefer shadow for motd and passwd changes
Busybox login has its own code for showing the motd and would have to
be patched for stateless Ostro to work as before. We use PAM for
interactive logins with ssh, so it seems better (because more
consistent) to also do the same for local logins, which can be
achieved by using login from shadow.

Busybox passwd also does not support creating /etc/shadow, while
shadow's passwd was already patched for that in Clear Linux.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>