swupd-client_2.87.bb: improve file downloading
Errors are now printed including the "staging" filename (and without
segfaulting, which was already fixed by the previous patch), and
download performance is a lot better on Ostro OS where IMA is active
and affects the performance of the close() syscall.

The download performance patch was also ported to current swupd
master, see https://github.com/clearlinux/swupd-client/pull/42. It
does not get included here because there is a chance to get it
via an upstream update, whereas that is unlikely for 2.87.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>