ostro-image-swupd.bb: add dbg and ptest packages to bundles
QA needs some way of running ptests on existing images, which implies
that the ptest packages need to be built together with the images. It
is not required that they are already pre-installed. The approach
taken here is to put them into a separate "world-ptest" bundle which
can be installed into both "ostro-image-swupd" and
"ostro-image-swupd-dev" via "swupd bundle-add".

The reason for choosing this approach is that normal users will not
need the ptests (thus including them by default is not necessary and
potentially confusing), and that having a second bundle is good for
testing swupd functionality.

Likewise, dbg packages were missing before. The usefulness of the
"ostro-image-swupd-dev" image is increased by including them directly
(better stack backtraces), so the "world-dev" bundle gets extended to
also include dbg packages. We cannot rename "world-dev" (swupd client
does not support bundle removal, see also
https://bugzilla.yoctoproject.org/show_bug.cgi?id=9493) and
"world-dev-dbg" is a bit awkward anyway, so the name is kept
unchanged.

The downside is the increased size of the "world-dev" bundle and thus
increased disk usage in "ostro-image-swupd-dev".

Fixes: IOTOS-1515

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>