ostro.conf: replace white/back-listing with complete recipe list
Choosing recipes via explicit black- and white-listing turned out to
be cumbersome and limited. On the one hand we blacklisted too much
(recipes desired by someone experimenting with Ostro OS as base were
blacklisted and thus unusable) and on the other hand not enough (our
world build still included components that we neither need nor
support, like apicd).

Now we maintain a simpler list of supported recipes and use the
supported-recipes.bbclass to configure our world build accordingly.

The current content of the ostro-supported-recipes.txt list is derived
from the build history of ostro-os master build #444 (= Ostro OS 1.0
with no new components added) and represents what was present in
actual images or bundles, i.e. without any component merely built
because the eSDK caused a world build. For example, acpid is now no
longer included in CI builds because it is not included in the list.

The ostro-supported.bbclass pre-build check is turned off because it
would confuse developers who just want to use Ostro OS with their own
custom recipes (the warning would trigger for those).

The intention is that only developers who add content to Ostro OS use
the warning. In the CI system itself it could be a fatal error, to
catch cases early where a submitter did not enable some required new
components first.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>