meta-swupd: swupd-image: improve how swupd inputs are staged
The previous process of creating images for each bundle which consisted
of the base image plus the bundles contents is resource intensive,
especially when a bundle consists of only a few files.

Instead of creating bundle images and performing several copies of the
entire mega-image rootfs, then pruning it down, we now make use of the
package manager to stage the files for non os-core bundles and generate
manifests based on the files installed by the package manager.

Once the mega image has been built and post-processed the contents
are re-copied from the mega rootfs to ensure that the final files
in the bundle 'chroot' match the files that will be used to populate
the image.

This patch refactors all of the logic for deploying the required inputs
for swupd (and the do_swupd_update task) into a single task, simplifying
the task scheduling for swupd-image significantly.

[YOCTO #9599]

(From meta-swupd rev: 714e77c917cd991e34fd8c7c53293cbf41e81563)

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>