meta-ostro: supported-recipes.bbclass: regex matching for recipe name and collection
Each entry in a supported recipes file is now treated as a pair of two
regular expressions, one for the recipe name (before the @) and one
for the collection name (after the @). This is backward compatible
with existing files.

The main purpose will be to enabled all recipes in the local workspace
(to be committed separately).

The entries get split up this way because it allows matching just by
recipe name or just by collection, which will be useful for providing
hints when there is no full match.

Shell wildcards could also have been used, but were not chosen because
of the extra flexibility provided by regular expressions.

(From meta-ostro rev: 541a2612b03d978cf70a77b20591d40f93261f28)

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>