meta-ostro: supported-recipes.bbclass: reformulate hint about the "supported elsewhere"
There is no check (and no way to add one) whether a recipe actually
exists in that other collection. It might have been removed there, or the
hint is based on a wildcard entry like ".*@workspacelayer".

In that case the message "only supported in" could be confusing if
understood as "exists and is supported there". Using "would be
supported in" makes it clear that this is a hypothetical situation and
only applicable if the recipe really can come from there.

(From meta-ostro rev: 64428f44accf8fc3839c703e13e9791b9f6d530e)

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>