ostro.conf: stay on gcc 5.3
Some of our components (iotivity, Edison kernel) do not compile with
gcc 6.1 yet. Until they do, we need to stay on gcc 5.3.

Without this override of the OE-core default we are stuck at the
OE-core master revision before c20d863da5.

Not the final commit, still need to file bugs.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>