meta-swupd: swupd/bundles: write manifest symlink for bundle manifests
When writing the rootfs.manifest for package manager staged bundles
also create a symlink to the manifest to better match the pattern
of manifests created for images.

We'll use this in a future commit to make the composed rootfs
construction in swupdimage based images more robust.

(From meta-swupd rev: dc663240eafab6ff0a3665470238ef77109319a4)

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>