meta-swupd: swupd-image: fetch and stage inputs for previous OS_VERSIONS
This change aims to help support more complex OS development
workflows where update artefacts:
* may not be sequential
* may have been generated on a separate system, or in a separate
  build directory

With this change we now generate a file after the swupd inputs have
been staged and an sstate object for the inputs created which maps
the OS_VERSION of the build to the name of the sstate object.

A new task will read in these 'map' files and try to ensure that
the swupd inputs are available before do_swupd_update, first by
checking for an sstate objects in SSTATE_DIR and when not present
attempting to fetch the object from an sstate mirror, before
unpacking the object into the swupd directory for processing.

[YOCTO #9321]

(From meta-swupd rev: ee51336e029f5e8dfb13f30b60c3e2a7ba786b3f)

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>