meta-openembedded: chrony: Versatile implementation of NTP
chrony is an alternative to ntpd.  In particular it may be useful
for quasi-realtime embedded systems that have a pulse-per-second
time reference available and need to rapidly synchronize to it after
boot, which appears to be unachievable with ntpd.

(From meta-openembedded rev: 9cd117a5dc5f26133a855d1cc3bf9604ae178685)

Signed-off-by: Henry Hallam <henry@pericynthion.org>
Reviewed-by: Koen Kooi <koen@dominion.thruhere.net>
Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>