swupdimage.bbclass: ensure that do_rootfs gets executed
OE-core commit 6d969bacc718e changed do_rootfs so that it creates
IMGDEPLOYDIR. That change broke the creation of additional swupd
images, because setting do_rootfs to empty caused the entire task to
be skipped, including the evaluation of the 'cleandirs' task
attribute.

It remains to be seen whether that's really the desired behavior, but
as it is what it is right now, we need to avoid the situation by
overwriting do_rootfs with non-empty code that doesn't do anything. That
way, the directory gets created.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>