ostroproject-ci.inc: avoid building ostro-image-swupd-all
That image got built and published although we don't want it because
do_populate_sdk_ext depends on do_build.

The original motivation ("make the eSDK as complete as possible") does
not really apply when the only difference in practice is one
additional QA bundle.

So now the eSDK gets built for ostro-image-swupd-dev, which has all
the relevant components and which gets built anyway.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>