meta-swupd: better performance with less reads from mega rootfs
Sharing pseudo instances is slow: using more than one avoids a
potential bottleneck (the pseudo daemon is often 100% busy on a CPU
during heavy use). Extracting files with full attributes also is
faster when merely sweeping through a tar archive, at least when most
of the content is needed.

This change increases performance by archiving the mega rootfs only
once and then copying or reading from that at those places where
previously access to the original rootfs and its pseudo instance was
needed.

TODO: occasionally, GNU tar fails with "Not found in archive" when
asked to extract a subset of the full.tar. Therefore not usable at the moment.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>