meta-swupd: switch to bsdtar
GNU tar had problems (a bug?!) extracting files based on a -T file
list: for some entries (like var/run) it complained with "Not found in
archive" although the entry was present. Wasn't easily reproducible
(for example, listing just that entry in the file list worked), so
instead of pursuing (and fixing) that problem, let's just switch to
bsdtar.

bsdtar is also nicer for other reasons:
- no need to explicitly add xattrs
- guaranteed to auto-detect compression, even when reading from
  stdin (GNU tar can only do that when working with files); not
  that relevant here, though
- uses less system calls when creating files, which should
  help a bit with performance under pseudo

TODO: this does not work! --no-recursion -T <file list> -xf is
needed for extracting a subset of the full.tar, but bsdtar does not
support that combination. Without --no-recursion, listing directories
extracts also their content, even when not listed.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>