linux-yocto: apply IMA setattr patches by default for Linux < 4.7
Maintaining a whitelist of kernel versions which need the setattr
patches just doesn't work: swupd broke in Ostro OS when switching to
kernel 4.4.14 (IOTOS-1712) and now again after switching to 4.4.20
(ostro-os/#199).

OTOH, not applying the patches by default enabled switching to 4.8
without having to do anything.

Now that the patches are known to be included in Linux 4.7, we can and
should apply the patches by default to all kernels older than that.

Fixes: ostro-os/#199

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>