libarchive: enable non-recursive extract/list
Required for meta-swupd performance enhancements: in meta-swupd, the
so called "mega" image contains a rootfs with all files that can
potentially be installed on a device. Other virtual image recipes need
a subset of those files or directories, and a partial extraction from
a single tar archive is faster than letting all virtual image recipes
share access to a directory under a single pseudo instance.

It may be necessary to extract a directory with all of its attributes
without the content of the directory, hence this patch.

TODO: decide about upstreaming it - upstream is willing to consider a
patch (see
https://groups.google.com/forum/#!topic/libarchive-discuss/JO3hqSaAVfs).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>