swupd-server: avoid pseudo xattr removal bug
Not removing the directories is okay: typically we don't
build incrementally, and we can remove any remaining ones before
invoking swupd.

Not removing a tempory directory tree may also have performance
benefits, but the even better solution will be to not write the tree
in the first place by calling libarchive directly.

Related-to: https://bugzilla.yoctoproject.org/show_bug.cgi?id=10623

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>