swupd server/client: update to client 3.7.2 and server 3.3.0
The two need to be updated in lockstep because of tool format changes,
leading to tools format 4.

Because of that, we need to keep the previous swupd-server around so
that we can generate intermediate versions for images with the
previous client.

This way, an Ostro OS 4400 (= 1.0) image can update to an intermediate
version with the new swupd client and from there to more recent
versions.

swupd client 3.7.2 has two of our patches merged
(0001-downloads-minimize-syscalls-to-improve-performance.patch,
0002-downloads-open-FILE-in-advance-and-use-default-write.patch) and
changes the HTTPS certificate checking so that the download uses the
system's trusted CAs instead of hard-coding a single one, so
Make-pinned-pubkey-configurable.patch is no longer needed. Furthermore,
the out-of-tree compilation was fixed upstream, so we can drop
the autotools-brokensep workaround.

swupd server 3.3.0 has merged
0001-delta.c-fix-xattr-test-after-patching.patch and
swupd_create_fullfiles-avoid-segfault-when-nothing-c.patch.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>