ostro.conf: revise swupd content reuse
The previous approach reused content from the latest build. The goal
was to make the build as fast as possible by maximizing reuse.

However, current Ostro OS release procedures do not guarantee that the
"builds" directory is kept around and therefore reusing content from
it is not currently possible. These procedures could be changed, but
until then we have to use the "milestone" directory.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>